vasnprintf: Avoid a dummy memory allocation.
* lib/vasnprintf.c (NOMEM_PTR): New macro.
(multiply, divide): Return it instead of NULL in case of memory
allocation failure.
(scale10_round_decimal_decoded): Update.

This backports a gnulib CL [1] to fix a -Walloc-size error that is
detected by the latest clang update.  Upstream bison has updated
their gnulib submodule to a version that includes this change, but
has not released a tarball that contains the updated files.

[1] https://gitweb.git.savannah.gnu.org/gitweb/?p=gnulib.git;a=commit;h=ffac79379d8b737473c67fdb92e278a325e96b68
Test: builds

Change-Id: I17f825950ff38f03e1e739027a754b8658f08721